Code review: Don't alias component and rename file ExtractNodes -> extractNodes to match method name
