+++ src/v2/Apps/Consign/Components/SoldRecently.tsx
@@ -4,7 +4,7 @@ import { QueryRenderer, createFragmentContainer, graphql } from "react-relay"
 import { useSystemContext } from "v2/Artsy"
 import { SoldRecentlyQuery } from "v2/__generated__/SoldRecentlyQuery.graphql"
 import { SoldRecently_targetSupply } from "v2/__generated__/SoldRecently_targetSupply.graphql"
-import { extractNodes } from "v2/Utils/ExtractNodes"
+import { extractNodes } from "v2/Utils/extractNodes"
 import FillwidthItem from "v2/Components/Artwork/FillwidthItem"
 import { Carousel } from "v2/Components/Carousel"
 import { SectionContainer } from "./SectionContainer"
