+++ src/v2/Utils/extractNodes.ts
